+++ src/IbanNet/Registry/Wikipedia/IbanWikipediaPattern.cs
@@ -1,0 +1,17 @@
+﻿namespace IbanNet.Registry.Wikipedia
+{
+    internal class IbanWikipediaPattern : WikipediaPattern
+    {
+        public IbanWikipediaPattern(string pattern)
+            : base(AdjustPattern(pattern)!)
+        {
+        }
+
+        private static string? AdjustPattern(string? pattern)
+        {
+            return pattern is null
+                ? null
+                : "2a,2n" + pattern;
+        }
+    }
+}